requirements.txt
deepgram-sdk>=3.0.0
pyaudio>=0.2.11
numpy>=1.21.0
pygame>=2.0.0
requests>=2.25.1
python-dotenv>=0.19.0
flask>=2.0.0
flask-socketio>=5.0.0
adafruit-circuitpython-ssd1306
Pillow